% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{build_http_env}
\alias{build_http_env}
\title{Build proxy url with authentification}
\usage{
build_http_env(proxy, username, password)
}
\arguments{
\item{proxy}{character. Proxy url}

\item{username}{character}

\item{password}{character}
}
\value{
proxy url with username and password
}
\description{
Build proxy url as \code{http://username:password@proxy-url.com:port}
}
\examples{
\dontrun{
  build_http_env("http://proxy-url.com:3232", "its", "me")
}
}
\keyword{internal}
