% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{check_proxy}
\alias{check_proxy}
\title{Check if proxy url is valid}
\usage{
check_proxy(proxy)
}
\arguments{
\item{proxy}{a character}
}
\value{
\code{TRUE} if proxy is valide
}
\description{
Checked against perl regex \code{"^http(s)?://[a-zA-Z0-9\\.-]+(:[0-9]+)?$"}
}
\examples{
check_proxy("http://10.132.23.444:3232")
}
