% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_proxy_activated}
\alias{is_proxy_activated}
\title{Verify if a proxy is configured}
\usage{
is_proxy_activated(verbose = FALSE)
}
\arguments{
\item{verbose}{if \code{TRUE} will print a summary of configuration. Authentification
is redacted for printing}
}
\value{
\code{TRUE} is a proxy is configured. \code{FALSE} otherwise.
}
\description{
This checks for any set environment variable relevant to proxy
}
\examples{
\dontrun{
  is_proxy_activated(verbose = TRUE)
}
}
