% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{read_password}
\alias{read_password}
\title{Ask interactively for authentification}
\usage{
read_password(prompt)
}
\arguments{
\item{prompt}{text prompt to user}
}
\value{
\emph{invisibly} the input of the user
}
\description{
This will use if available in this order
\itemize{
\item RStudio IDE prompt
\item Console input
}
}
\examples{
if (interactive()) {
 pwd <- read_password("password")
}
}
