% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{unset_proxy}
\alias{unset_proxy}
\title{Unset proxy configuration for the session}
\usage{
unset_proxy(verbose = FALSE)
}
\arguments{
\item{verbose}{default is \code{FALSE}.}
}
\value{
\code{FALSE} is no proxy configuration was set before. \code{TRUE} if
configuration is unset with success.
}
\description{
This fonction is a helper to unset all configuration of proxy for the current
session.
}
\examples{
 set_proxy(proxy = "http://10.132.23.444:3232",
           username = "",
           password = "",
           noproxy = ".mycompany.com",
           https = TRUE)
 is_proxy_activated(verbose = TRUE)
 unset_proxy(verbose = TRUE)
 is_proxy_activated(verbose = TRUE)
}
