% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_proxy}
\alias{set_proxy}
\title{Configure the proxy for internet connection}
\usage{
set_proxy(proxy = getOption("proxyconfig.proxy", default = NA_character_),
  username = NULL, password = NULL, noproxy = NULL, https = TRUE)
}
\arguments{
\item{proxy}{a character giving the proxy url. By default, it will fetch for
option \code{proxyconfig.proxy} that can be set accross sessions in
\emph{.Rprofile} or \emph{Rprofile.site}.}

\item{username}{character. If \code{NULL}, user will be prompted}

\item{password}{character. If \code{NULL}, user will be prompted}

\item{noproxy}{character vector of domain that proxy should ignore}

\item{https}{logical. If \code{TRUE} \emph{HTTPS_PROXY} will be used}
}
\value{
\code{TRUE} \emph{invisibly} if no error. Side effects are the environment
variables for proxy being set for the session. You can check them with
\code{\link[base:Sys.getenv]{base::Sys.getenv()}}

\code{FALSE} \emph{invisibly} with a warning if a proxy configuration was already set. You need
to unset the configuration with \code{\link[=unset_proxy]{unset_proxy()}} before setting a new one.
}
\description{
this will help to set up the proxy configuration using
environment variable
\itemize{
\item url of the proxy with port
\item authentification if necessary (username and password)
\item any domain that proxy should ignore
\item activate or not https proxy
}

You can set a proxy url across sessions by setting option \code{"proxyconfig.proxy"}.
}
\section{Proxy Environment Variable}{


The proxy is set using environment variables
\itemize{
\item \code{HTTP_PROXY}
\item \code{HTTPS_PROXY}
\item \code{NO_PROXY}
The lower case version are also set because some utilities do not understand
upper case environment variable.
}

This configuration is used by \code{curl} for internet connection and other
\emph{method}. See \emph{Setting Proxies} in \code{\link[utils:download.file]{utils::download.file()}}
}

\examples{
\dontrun{
 set_proxy(proxy = "http://10.132.23.444:3232",
           username = "",
           password = "",
           noproxy = ".mycompany.com",
           https = TRUE)
is_proxy_activated(verbose = TRUE)
}
}
